Remove the lock before the job is peformed
If another job is queued after perform is finished, but before the lock is removed, it will be dropped. Consider if this job looks up an object and does something depending on the sets a value based on the current state of the object. If another job is queued because the object changed again between when the object is looked up and when the lock is removed, it will not have the value it sets is no longer accurate, but the job won't be queued again to update it.